git-lfs: Add xcodegit variant that depends on Xcode git
git-lfs depends on git, which has 60 (recursive) dependencies. However,
many MacPorts users already have Xcode Command Line Tools installed,
which comes with git. The xcodegit variant uses this included version of
git if it is available, reducing the amount of non-build dependencies
that are potentially installed from 61 to 0.

While file-based dependencies are unusual in MacPorts, the MacPorts
Guide (Section 5.4.1) says that an example of when they can be used is
"with ubiquitous utilities...where the versions in macOS are often
sufficient". Since git-lfs has a minimum requirement of Git 1.8.5
(October 2013), macOS's version (2.39.3, April 2023) is more than
sufficient.